all tests passed, create dockerfile
