+++ README.md
@@ -3,3 +3,18 @@
 В директории `tests` находятся тесты для проверки API, которое должно быть реализовано в веб-сервере.
 
 Директория `web` содержит файлы фронтенда.
+
+Для работы Web-сервиса необходимо создать .env файл в корне проекта со следующими параметрами:
+```
+SERVER_PORT=Ваш порт для запуска приложения(по умолчанию 7540)
+TODO_DBFILE=Имя базы данных(по умолчанию scheduler.db)
+TODO_PASSWORD=Пароль для доступа к веб сервису
+```
+Существующую базу данных необходимо поместить в директорию ```db```
+
+Для запуска контейнера выполните команды ниже(при необходимости заменти на нужные вам параметры)
+
+```
+docker build -t scheduler-app .
+docker run --name scheduler-container -p 7545:7545 -v /path/to/database/in/host:/myapp/db -d scheduler-app
+```
