+++ internal/api/server.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 	"net/http"
 	"time"
+
+	myLog "github.com/subbbbbaru/first-sample/pkg/log"
 )
 
 type Server struct {
@@ -18,9 +20,6 @@ func (s *Server) Run(port int, handler http.Handler) error {
 		ReadTimeout:    10 * time.Second,
 		WriteTimeout:   10 * time.Second,
 	}
+	myLog.Info().Printf("Server run at %s port\n", s.httpServer.Addr)
 	return s.httpServer.ListenAndServe()
 }
-
-// func (s *Server) Shutdown(ctx context.Context) error {
-// 	return s.httpServer.Shutdown(ctx)
-// }
