+++ internal/handlers/createtask.go
@@ -5,9 +5,10 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/subbbbbaru/first-sample/pkg/log"
 	"github.com/subbbbbaru/go_final_project/internal/models"
-	"github.com/subbbbbaru/go_final_project/internal/repository"
 	"github.com/subbbbbaru/go_final_project/internal/service"
+	"github.com/subbbbbaru/go_final_project/utils"
 )
 
 type Handler struct {
@@ -29,6 +30,7 @@ func (h *Handler) TaskHandler(w http.ResponseWriter, r *http.Request) {
 	case http.MethodDelete:
 		h.DeleteTaskHandler(w, r)
 	default:
+		log.Info().Println("method not yet realized")
 		w.WriteHeader(http.StatusBadRequest)
 		json.NewEncoder(w).Encode(map[string]string{"error": "method not yet realized"})
 		return
@@ -40,18 +42,20 @@ func (h *Handler) CrateTaskHandler(w http.ResponseWriter, r *http.Request) {
 	var task models.Task
 	err := json.NewDecoder(r.Body).Decode(&task)
 	if err != nil {
+		log.Error().Println(err)
 		w.WriteHeader(http.StatusBadRequest)
 		json.NewEncoder(w).Encode(map[string]string{"error": err.Error()})
 		return
 	}
 
 	if task.Title == "" {
+		log.Error().Println("task title not found")
 		w.WriteHeader(http.StatusBadRequest)
 		json.NewEncoder(w).Encode(map[string]string{"error": "task title not found"})
 		return
 	}
 
-	timeNow := time.Now().Truncate(24 * time.Hour).UTC()
+	timeNow := time.Now() //.Truncate(24 * time.Hour)
 
 	if task.Date == "" {
 		task.Date = timeNow.Format("20060102")
@@ -59,27 +63,30 @@ func (h *Handler) CrateTaskHandler(w http.ResponseWriter, r *http.Request) {
 
 	date, err := time.Parse("20060102", task.Date)
 	if err != nil {
+		log.Error().Println(err)
 		w.WriteHeader(http.StatusBadRequest)
 		json.NewEncoder(w).Encode(map[string]string{"error": err.Error()})
 		return
 	}
 
-	repos := repository.NewRepository(nil)
 	if date.Before(timeNow) {
 		if task.Repeat == "" {
 			task.Date = timeNow.Format("20060102")
 		} else {
-			task.Date, err = repos.NextDate(timeNow, task.Date, task.Repeat)
+			nextDate, err := utils.NextDate(timeNow, task.Date, task.Repeat)
 			if err != nil {
+				log.Error().Println(err)
 				w.WriteHeader(http.StatusBadRequest)
 				json.NewEncoder(w).Encode(map[string]string{"error": err.Error()})
 				return
 			}
+			task.Date = nextDate
 		}
 	}
 
 	id, err := h.services.TodoTask.Create(task)
 	if err != nil {
+		log.Error().Println(err)
 		w.WriteHeader(http.StatusBadRequest)
 		json.NewEncoder(w).Encode(map[string]string{"error": err.Error()})
 		return
@@ -89,12 +96,6 @@ func (h *Handler) CrateTaskHandler(w http.ResponseWriter, r *http.Request) {
 		ID int64 `json:"id"`
 	}{ID: id}
 
-	// jsonee, err := json.Marshal(response)
-	// if err != nil {
-	// 	w.WriteHeader(http.StatusBadRequest)
-	// 	json.NewEncoder(w).Encode(map[string]string{"error": err.Error()})
-	// 	return
-	// }
 	w.Header().Set("Content-type", "application-json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(response)
