+++ internal/handlers/deletetask.go
@@ -2,39 +2,43 @@ package handlers
 
 import (
 	"encoding/json"
-	"log"
+
 	"net/http"
 	"strconv"
+
+	"github.com/subbbbbaru/first-sample/pkg/log"
 )
 
 func (h *Handler) DeleteTaskHandler(w http.ResponseWriter, r *http.Request) {
 	id := r.URL.Query().Get("id")
 
 	if len(id) == 0 {
+		log.Error().Println("wrong id")
 		w.WriteHeader(http.StatusBadRequest)
 		json.NewEncoder(w).Encode(map[string]string{"error": "wrong id"})
 		return
 	}
 	taskId, err := strconv.Atoi(id)
 	if err != nil {
+		log.Error().Println(err)
 		w.WriteHeader(http.StatusInternalServerError)
 		json.NewEncoder(w).Encode(map[string]string{"error": err.Error()})
 		return
 	}
 
-	task, err := h.services.TodoTask.Delete(taskId)
+	_, err = h.services.TodoTask.Delete(taskId)
 	if err != nil {
+		log.Error().Println(err)
 		w.WriteHeader(http.StatusInternalServerError)
 		json.NewEncoder(w).Encode(map[string]string{"error": err.Error()})
 		return
 	}
-
-	log.Println("TASK DELETE = ", task)
 
 	w.Header().Set("Content-type", "application-json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 	err = json.NewEncoder(w).Encode(map[string]string{})
 	if err != nil {
+		log.Error().Println(err)
 		w.WriteHeader(http.StatusInternalServerError)
 		json.NewEncoder(w).Encode(map[string]string{"error": err.Error()})
 		return
