+++ internal/handlers/gettaskbyid.go
@@ -4,17 +4,21 @@ import (
 	"encoding/json"
 	"net/http"
 	"strconv"
+
+	"github.com/subbbbbaru/first-sample/pkg/log"
 )
 
 func (h *Handler) GetTaskByIdHandler(w http.ResponseWriter, r *http.Request) {
 	id := r.URL.Query().Get("id")
 	if len(id) == 0 {
+		log.Error().Println("wrong id")
 		w.WriteHeader(http.StatusBadRequest)
 		json.NewEncoder(w).Encode(map[string]string{"error": "wrong id"})
 		return
 	}
 	taskId, err := strconv.Atoi(id)
 	if err != nil {
+		log.Error().Println(err)
 		w.WriteHeader(http.StatusInternalServerError)
 		json.NewEncoder(w).Encode(map[string]string{"error": err.Error()})
 		return
@@ -22,6 +26,7 @@ func (h *Handler) GetTaskByIdHandler(w http.ResponseWriter, r *http.Request) {
 
 	task, err := h.services.TodoTask.GetTaskById(taskId)
 	if err != nil {
+		log.Error().Println(err)
 		w.WriteHeader(http.StatusInternalServerError)
 		json.NewEncoder(w).Encode(map[string]string{"error": err.Error()})
 		return
@@ -31,6 +36,7 @@ func (h *Handler) GetTaskByIdHandler(w http.ResponseWriter, r *http.Request) {
 	w.WriteHeader(http.StatusOK)
 	err = json.NewEncoder(w).Encode(task)
 	if err != nil {
+		log.Error().Println(err)
 		w.WriteHeader(http.StatusInternalServerError)
 		json.NewEncoder(w).Encode(map[string]string{"error": err.Error()})
 		return
