+++ internal/models/task.go
@@ -1,7 +1,6 @@
 package models
 
 type Task struct {
-	// ID      int64  `json:"id" db:"id"`
 	ID      string `json:"id" db:"id"`
 	Title   string `json:"title" db:"title"`
 	Date    string `json:"date" db:"date"`
