+++ tests/settings.go
@@ -1,7 +1,7 @@
 package tests
 
-var Port = 7540
-var DBFile = "../scheduler.db"
+var Port = 7545
+var DBFile = "../db/scheduler.db"
 var FullNextDate = true
 var Search = true
 var Token = ``
